reorganize modules into package; add .gitignore file
