+++ run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+export PYTHONPATH=janggi:${PYTHONPATH}
+./janggi/gui.py
