+++ test.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+export PYTHONPATH=janggi:${PYTHONPATH}
+python -m unittest discover -v -s janggi/
